# framework-depth-lab/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Framework Depth Lab Items API
  version: 1.0.0
  description: Identical minimal contract used by the Spring Boot, FastAPI, and Gin examples.
paths:
  /healthz:
    get:
      operationId: health
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
  /items/{item_id}:
    get:
      operationId: getItem
      parameters:
        - name: item_id
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Item
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Item" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Problem" }
  /items:
    post:
      operationId: createItem
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateItem" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Item" }
        "422":
          description: Validation failure
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Problem" }
components:
  schemas:
    Health:
      type: object
      required: [status]
      properties:
        status: { type: string, const: ok }
    CreateItem:
      type: object
      additionalProperties: false
      required: [name, price]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        price: { type: number, format: double, minimum: 0 }
    Item:
      allOf:
        - $ref: "#/components/schemas/CreateItem"
        - type: object
          required: [id]
          properties:
            id: { type: integer, minimum: 1 }
    Problem:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }